refactor: imgaecontroller
